+++ app/src/pages/ProfileGeneralPage/ProfileGeneralPage.tsx
@@ -1,5 +1,5 @@
 import { Dashboard } from '@/components/templates/Dashboard';
-import { useProfileGeneralPageDashboard } from './hooks/useProfileGeneralPageDashboard';
+import { useProfileGeneralPageDashboard } from './hooks';
 
 export const ProfileGeneralPage = () => {
   return <Dashboard {...useProfileGeneralPageDashboard()} />;
